requirements.txt
fastdownload
tqdm
transformers 
pillow
diffusers
matplotlib
